refactor(database): drop redundant raw include shim
