--- src/middleware/database/raw.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "raw/raw.hpp"
